Tuning map to load passability and events. Tuning frame to take the place of flipping (which is frame based). Cleanup in map on some context issues.

+++ include/Game/Map/Tile.h
@@ -214,8 +214,8 @@ public:
   void setY(int16_t y);
   
   /* Updates the relevent enter and exit events, from file data */
-  bool updateEnterEvent(XmlData data, int file_index, int section_index);
-  bool updateExitEvent(XmlData data, int file_index, int section_index);
+  bool updateEnterEvent(XmlData data, int file_index, uint16_t section_index);
+  bool updateExitEvent(XmlData data, int file_index, uint16_t section_index);
   
   /* Unsets the base layer(s) */
   void unsetBase();
